move OutputSpan-providing extensions to their own file
- and their tests, too
