+++ Tests/FutureTests/OutputSpanTests.swift
@@ -11,7 +11,6 @@
 //===----------------------------------------------------------------------===//
 
 import XCTest
-import Foundation
 import Future
 
 struct Allocation<T>: ~Copyable {
@@ -211,43 +210,5 @@ final class OutputBufferTests: XCTestCase {
     catch MyTestError.error {
       XCTAssertEqual(a.isInitialized, false)
     }
-  }
-}
-
-final class OutputBufferUsageTests: XCTestCase {
-
-  func testArrayInitializationExample() {
-    var array: [UInt8]
-    array = Array(capacity: 32, initializingWith: { output in
-      for i in 0..<(output.capacity/2) {
-        output.appendElement(UInt8(clamping: i))
-      }
-    })
-    XCTAssertEqual(array.count, 16)
-    XCTAssert(array.elementsEqual(0..<16))
-    XCTAssertGreaterThanOrEqual(array.capacity, 32)
-  }
-
-  func testDataInitializationExample() {
-    var data: Data
-    data = Data(capacity: 32, initializingWith: { output in
-      for i in 0..<(output.capacity/2) {
-        output.appendElement(UInt8(clamping: i))
-      }
-    })
-    XCTAssertEqual(data.count, 16)
-    XCTAssert(data.elementsEqual(0..<16))
-  }
-
-  func testStringInitializationExample() {
-    var string: String
-    let c = UInt8(ascii: "A")
-    string = String(utf8Capacity: 32, initializingWith: { output in
-      for i in 0..<(output.capacity/2) {
-        output.appendElement(c + UInt8(clamping: i))
-      }
-    })
-    XCTAssertEqual(string.utf8.count, 16)
-    XCTAssert(string.utf8.elementsEqual(c..<(c+16)))
   }
 }
